The string which separates a hint from the rest of the contents of a cloze is now customisable by setting the variable 'org-drill-hint-separator'. The string is '||' by default (it was previously a single '|').
New global variable 'drill-answer'. If this variable is set to a
non-nil value, then the default 'show answer' behaviour is to
display the value of the variable instead of the contents of the
drill item. This can be useful if you make custom drill card types
which randomly generate problems (e.g. random maths problems, or
translating a random number to or from another language.) By setting
this variable to the answer when you generate the question, you
can avoid having to write a custom 'show answer' function.

Fixed a bug where after suspending the drill session via 'q'uit
or 'e'dit, the session could not be resumed if there was only one
item left to test.